Use java code for validation
This is almost the java code that we use for validation on the
wps wrapper - but now as standalone jar file with CLI interface.

I also extended the tasks.py for invoke in order to make the
validation with this jar file.

In the meantime Hugo merged my PR which sets the seed in a position
that the file is valid according to the schema.